fix: adjust nested stats test
