+++ datafusion/datasource/src/mod.rs
@@ -38,7 +38,6 @@ pub mod file_scan_config;
 pub mod file_sink_config;
 pub mod file_stream;
 pub mod memory;
-pub mod nested_schema_adapter;
 pub mod schema_adapter;
 pub mod sink;
 pub mod source;
